Improve the icon parsing for files with a slightly wrong BMP header.
The reason it failed was that we always expected the height property
of the BMP header to be double the height of the icon. The
kde_favicon.ico did not fulfill this requirement.  We can fix that by
simply reading from the ICONDIR entry instead, since that has always
the correct height.

Task-number: 229829
Reviewed-by: alexis