Add qcore_unix_p.h containing mostly safe versions of Unix functions.
Most of these functions are from unistd.h and need to have a loop
around the actual call because the calls can be interrupted by a
signal delivery.

Some special calls (open, dup2, pipe) require an extra flag to support
thread-safe execution: the file descriptor must be created from the
operating system with the FD_CLOEXEC flag already set.

The O_CLOEXEC flag is specified in POSIX.1-2008, but the rest is
Linux-specific.

Reviewed-By: João Abecasis