Continuing tutorial
Start of qdoc componenet relating to example files.
Also removed t8 and replaced with Final, which will represent
the final 'application', and not necissarily in terms of
minor stages.  However difference that will be useful or
confusing to the user will still be highlighted, just not
in the progressive manor of the rest of the tutorial

Scheduled for review when complete.