Implement QRasterPaintEngine::clip(QRegion) to do something sensible...
Since we now store QRegion's in the clipdata is possible for us to
detect that we should go back to a rect-clip mode when we get to
clip(region) or clip(rect)

It turns out that the qt_span_fill_clipRegion is very slow, even for
smaller regions, testcase was 3 rects, its 10% worse than span based
clipping. For a fullscreen ellipse it was more in the range of 100x
worse, so I removed the clip function entirely.