Carbon/Cocoa: It is possible to resize a widget above its maximum size
This is not strange since we never did anything to limit a resize within
the max min boundries. This patch factores out the code that ensures
this into a private function that is called both as a reaction to a
resize event, but also if resize is done programatically.

Task-number: 251893
Reviewed-by: Trenton Schulz