Allow QThread to start a QTimer and QEventLoop before QCoreApplication is constructed
Fix a regression from 4.3 to 4.4 that prevented QThread from starting
timers and event loops before QCoreApplication was instantiated. Even
though this is documented not to work, it seems that people have been
relying on it.

This reverts commit e52e5acdfa198cd079bbfe3a9302debf46c7cadd which
attempted to work around not calling g_thread_init() multiple
times. The proper fix is to serialize the g_thread_supported() checks
in the Glib event dispatcher.

Task-number: 248191
Reviewed-by: denis