Fixes for QByteArrayMatcher
Copy constructor and assignment operator lose data: pointer to content
and the length of content also need to be copied over.

QByteArrayMatcher::pattern() would return a null byte array if instance
was initialized with c-string.

Changed default constructor to explicitly initialize pattern length to
zero.

The bug in the assignment operator is a regression against 4.4.3.

Task-number: 251958
Reviewed-by: MariusSO
Reviewed-by: paul