Remove the simpleBrush concept
Since we have to update the brush for each paint command anyway keeping
the state of the brush and caching the value of simpleBrush makes no
sense.

Also, most times when filling with a brush
QPaintEngineEx::fillRect(const QRect &, const QBrush &) is called in
which we never got a brushChanged() anyway.

Reviewed-by: Donald <qt-info@nokia.com>