Fix background painting.
This is a reverted patch that has been rebased on to another branch and
then caused conflicts so it might contain an error or two. Basically we
need to store the background texture from the style into the palette so
that it can be replaced by applications that don't want it.