Be able to set a border-image on plain QWidget
we were not able to set border-image on plain QWidget,
 QDialog, QMainWindow, QFrame, QMenu, and more. And for some,
not even border was working.

PE_Widget used to only draw the background,  now it draws both the whole
rule (background + border)

Since menu now have a PanelMenu, let's use that insead of PE_Widget +
PE_FrameMenu.  As for QFrame, we do not need to draw the border in
CE_ShapedFrame, as PE_Widget does it.

This patch mostly move code arounds.

Reviewed-by: Jens Bache-Wiig