QPixmapCache:Remove the old pixmap if you insert one with the same key.
If you insert a new pixmap in the cache with a key that was already in
the cache then we remove the old pixmap and add the new one. This avoid
to fill the memory with garbage even if the cache has a protection to
avoid running out of memory. This was discovered with QraphicsView
and its cache. We don't need to keep old cached pixmaps for an item.

Task-number: KDE
Reviewed-by: Trond