Cocoa: qwidget auto test fails (windowMoveResize)
The test fails because, in cocoa, when resizing a window to zero (either
w or h), cocoa will also move the window up in the corner (!). So the
fix is to issue an extra move back to it's true location after the
resize. The faulty cocoa move is issued inside the resize callback, so
we choose to not update the window location anymore from a pure resize
callback.

Task-number: 251895
Reviewed-by: Trenton Schulz