Made tabbar tab icons set the On state for selected tabs
This makes it possible to make a custom icon for the selected tab.
We cannot use the Selected state because that has a blue tint on
many platforms which is not what you typically want int a tab. The
On state will only affect icons where these states are explicitly set.

Task-number: 202063
Reviewed-by: ogoffart