Undo implementation of setWindowOpacity_sys().
This function was implemented using API from Symbian that will be
deprecated. It was never actually suitable anyway since this function
cannot be called after the window has been shown and this needs to be a
runtime decision for Qt. The solution is to  use
SetTransparencyAlphaChannel(), but that will come later.