Fixed temporary file rename problem in Symbian OS.
In Symbian OS temp file rename will always go to fallback solution
since close for temp file does not really close the handle. The fallback
implements rename by copy and remove. But in symbian OS also the remove
after copy failed since it was done through local 'in' object. However
the file was still open in object doing the remove.

The fix is to close the local 'in' object and use current 'this' object to
remove the file. This object points to QTemporaryFile and it will make sure
that temp file is first closed before trying to remove.