Fix focus issues in QX11EmbedContainer, affecting Flash plugin in QtWebKit.
We are now moving focus to the XEmbed focus proxy when we accept a
client and when we get window activation. Window activation did that
before, but only when we were the active container, thus had focus in.
Now we do it always.

Due to race conditions with the window manager, the time stamt we used
for XSetInputFocus was the same as that of the window manager. This
broke it from time to time in Metacity and Xfce but always in KWin.
With other tested window manager we didn't have this issue.

Following Owen Tayler's advice (one of the authors of the XEmbed
specification) we now use CurrentTime and not qt_x11Data->time
when moving the input focus as there is no explicit user interaction
involved.

Reviewed-by: Denis Dzyubenko <denis.dzyubenko@nokia.com>
Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>