QApplication::setStyle() can cause a crash
QApplication::setStyle() caused a crash if called before
constructing the QApplication instance for custom styles. The reason
was that polish tried to create a pixmap (which is not allowed before
qApp is running). This fix checks that qApp exists. Polish will anyway
be called again when qApp gets constructed.

Task-number: 243697
Reviewed-by: Bjørn Erik Nilsen