Avoid overwriting a user set palette.
The polish(QWidget*) function is called the first time a widget is
shown and therefore setting the widget's palette in this function will
overwrite the user set palette if one is specified before show() is
called. The solution is to break this into two parts, the first part
sets the global palette in polish(QApplication*) and this is the
part of the palette that is independent of widget. The second part
involves adding the widget specific palettes using the QApplication::
setPalette() function that takes a classname and this is done in the
polish(QWidget*) function.