Fixed busy indicator for a QProgressBar with a style sheet applied to it.
For a progress bar with a style sheet applied to it, this
fix ensures that the timer event is passed to the event
handler that updates the busy indicator animation state.
Essentially, the bug was that the decision that the
event was processed by the proxy style object (baseStyle())
was based only on the return value from the event() function.
In this case it is necessary to check that the event was
accepted as well.

Reviewed-by: ogoffart
Reviewed-by: brad
Task-number: 252283