QTextEdit::ExtraSelection failure with style sheets
the feature has to handle text with and without background, and extra
selections with and without background, or even only with underline style.
Trouble is that you sometimes want to accumulate styles, for example spell
checking wiggly underline plus search result highlights or background markup
from the css stylesheet.

Task-number: 252310