Ensure that we send the Apple Events even when Cocoa isn't ready.
In general, Cocoa handles the the Apple Events for us. However, this is
time between creating the NSApplication and Cocoa has set everything up,
usually after the event loop is running. This means that until that
time, the events are dropped on the floor :-/. The workaround is to use
the same handler that we use for Carbon, but to only have it enabled for
until Cocoa is ready to handle things. This will result in  not stepping
on the toes when used in a plugin (if it does, we can conditionalize
it).

Task-number: 252795
Reviewed-by: Richard Moe Gustavsen