Fix regression in drawing parts of pixmaps on Panther.
The width of a pixmap is NOT the same as the bytes per row. We may have
been able to do that in the past, but it doesn't work in the general
case. I also took the liberty to introduce a QCFType to automatically
handling the free in the general case.

Task-number: 252176
Reviewed-by: sroedal