Fix OpenSSL dynamic loading on OpenBSD.
OpenBSD's OpenSSL libraries are linked in a bizarre way: libssl.so
doesn't link to libcrypto.so, even though it depends on it. I don't
claim to understand why, but they do it. So make sure we export its
symbols for libssl to see and we load libcrypto first.

Task-number: 252042
Patch by: Marc Espie <espie@openbsd.org>
Reviewed-by: Peter Hartmann