Fixed qDebugs within tests being written as "error" tags in the xunitxml testlog.
`error' in xunit causes many xunit processing tools to consider the test
as failing.  This is different semantics from native testlib XML, where
a qDebug is considered a message and not a failure.

Change it to put qDebugs under the "system-err" tag when using xunitxml
to retain semantic compatibility with testlib XML.

Autotest: tst_selftests