KeyProxy to send to focused child of focus realms.
If a component is loaded into a focus realm, it is likely when
specifying it as a target of a focus proxy the intent is to
have it behave as if the focus proxy target had focus.

This means for focus scopes, descend to find the focused descendent
of that scope, rather than sending directly to the item itself.