qdoc: Moved platform-specific qdoc comments to common .cpp file.
When building docs for the mac, qdoc comments for functions defined
in the .h file were not found in any of the .cpp files in the mac
package because they were in the x11 or windows .cpp file. So I
moved them to a .cpp file that is in all the packages.

#Task-number: 252496