Make sure machine enters error state if history state has no default state
Keep searching the parent hierarchy for error states even if a state in the
hierarchy cannot be cast to QState. Also make currentErrorState==0 an assert,
since there should always be an error state (we default to the special
initialErrorState if we are unable to find anything else), otherwise the
machine might get into an undefined state (e.g. configuration is empty)