Win: QLineEdit not updating correctly when embedded in Graphics View
Updates triggered by the line edit itself, i.e. cursor blinking, are not
processed after the top-level is resized. This is yet another problem
caused by the event dispatcher on Windows (Qt posted events are not sent
during top-level resize, task 146849). We added a work-around for that
particular case by posting an event via Windows, but the widget is not
visible on the screen (hidden from Windows' POV) so it'll never be posted.
And of course then we'll never receive it and the backing store is not
synced. This work-around is therefore useless for widgets that are
not visible on the screen.

However, not receiving update requests while resizing the top-level
(in this case QGraphicsView), is not a problem for embedded widgets
because all items and hence the proxied widgets are repainted by
graphics view anyways.

Task-number: 252400
Reviewed-by: Olivier