Re-enabled tst_QUdpSocket::performance with proxy for Symbian OS.
It was thought this test case was hanging earlier with Socks5Proxy,
but actually it was not hanging but just taking a lot of time to
complete. The reason for slowness was UDP packet fragmentation.

Without further studies I decided to decrease the written buffer size
for Symbian OS when using Socks5Proxy. This seems to solve the problems.

Also added some debug print why performance in Symbian OS is 0.00MB/s
without proxy.