Add a  AA_MacDontSwapCtrlAndMeta application attribute.
This is to help undo the some magic that is in the Qt/Mac port. Qt
automatically flips the Meta and Control keys on Mac. This is a
"feature" that makes porting older programs that don't use standard
shortcuts easier as Ctrl and Command usually map to the same shortcuts
in the application. The upshot of this is that I need to strip the
text() out of key events if they contain the Control or Meta modifier.
This causes much headache for anyone writing a terminal emulator. Though
they would still have to write special code because the keys are swapped
anyway. This allows people to write the terminal emulator where hitting
the Control key will really send a Control key modifier.

We've also done the extra work to ensure that standard shortcuts work
correctly regardless of what the value of the attribute is. That is, if
you specify QKeySequence::Cut for a shortcut you can always hit
Command+X and things will work.