Removed CActiveScheduler::Install(NULL) from closingDown.
When calling QThread::terminate - closingDown is not called from
QThread context, but from thread which is called terminate.

In this case, if we Install(NULL) AS, we will actually remove the AS
from calling thread not from terminated thread.

Since thread is anyway going to be terminated there is not much point
on setting AS to NULL.