Improved gesture propagation.
Each gesture can now be accepted separately and not accepted gestures
will be propagated to parent widget that are subscribed to them.

Added an internal flag to specify that gesture is a "singleshot" - aka
if the gesture is not continious and will not have a GestureStarted
state, but only GestureFinished.

Asynchronous gestures still need to fixed, as well as QGraphicsView.