Add the ARMv6 inline assembly code for compiling with RVCT.
This is basically a copy & paste of the GCC inline assembly above,
switched to the RVCT inline assembly model (which is actually easier
to write and understand).

I verified that this code compiles and assembles as expected. The
output generated by RVCT is pretty much on the mark. However, I have
not executed this code yet to see if it performs as expected.

To be noted:
 - when expanding the inline template code, RVCT may be tempted to
   switch your entire function to ARM mode. Should we add
   __attribute__((noinline)) to prevent that?
 - There's no equivalent to GCC inline assembler's clobber, especially
   of "memory". Also, there's no "volatile" qualifier to the
   assembly. Does the compiler know it can't reorder the code? Does it
   know it shouldn't trust the value of the memory after this? My test
   indicates the code is fine...

Reviewed-By: Shane Kearns
(cherry picked from commit 43c607383697ab506f9eb0d491ec6348f939e53d)