Faster slot disconnection while destroying of QObject
Destroying a QObject which is connected to many signals was slow.
We had to loop over all the ConnectionsList of all the connected object
to remove the connections. (in QObjectPrivate::removeReceiver)

The idea is to now use lists of pointer to Connection both in the
connection list, and in the list of senders. So clearing the connection
should be faster.

Task-number: 251097
Reviewed-by: Brad