Improved caching of QSvgIconEngine::actualSize() calls.
If an application calls QIcon::actualSize() for SVG based icons, before
the icon is drawn for the first time, the actualSize() call will
result in the SVG file being loaded and parsed for each call.
Instead of doing that, just render a pixmap of the queried size and
then the resulting actualSize() calls will used the size of the
cached pixmap.

Task-number: related to 251106
Reviewed-by: Kim