Add the chart utility from qtestlib-tools.
It is now possible run the benchmarks with "-chart" to generate
charts for benchmarks that produce chartable data. See
examples/qtestlib/tutorial5 for an example.

This patch has minimal inpact on QTestLib itself. Most of the
functionality is implemented in a separate tool which QTestLIb
calls after detecting the chart command line option.