Don't interrupt test before animation finishes
We're testing if both animations actually run and finish, so we need to make
sure one of the polished signals is emitted last, and then react to this. I've
done this by setting the duration of the animation which animated the property
set by s2Child, so that it's sufficient to listen to this polished signal.