Makes the layout of many of the areas be in aligned columns to immensely increase readability
Unfortunately the patch causes quite some regressions and I think some refectoring of this class
would be needed.
I didn't do that since there is no API docs, the variable naming is not very clarifying and the
code is hard to read due to it not using the coding style and mixing tabs as well as spaces.