Fixed the translation from S60 text format to Qt text format in FEP.
This was done by making sure that the format code is called from both
start and update, because S60 will often return nothing in start, but
then later on return something in update.

In addition, we fall back on our own format if S60 does not give us
anything. Currently it seems to give us a format only when using
predictive text.