Implement QEvent::GraphicsSceneTouchBegin propagation
This event will be propagated to all items under the first touch point,
otherwise it's the same as QEvent::TouchBegin propagation: the first
item to accept the event gets the remaining touch events. If none of
the GraphicsSceneTouchEvents sent to the items are sent, then the
GraphicsSceneTouchEvent that was sent to the seen will be ignored,
causing the QTouchEvent sent to the QGraphicsView to also be ignored,
meaning widget TouchBegin event propagation resumes.