qdoc: Ensure that all inner class methods are included in the function index.
Since the original code only used the parent class name as a key in the
map that collects functions, you would only ever see either
QHash::iterator::key() or QMap::iterator::key() in the index. This patch
changes the key to the fully-qualified name for the parent class so that
inner class methods are included.

Reviewed-by: Martin Smith <msmith@trolltech.com>