Fixed: No margin between the text and the shortcut in menu with stylesheet
If padding, margin border are set on an menu item with stylesheet,
no space appeared between the text and the shortcut.

Fix that by adding the 12px space, same as in QCommonStyle equivalent
function

Note that this patch also remove unused code, since the variable with
and height were not used.

example to reproduce the bug: the following stylesheet on a mainwindow
that contains a menu with one action with a shortcut
  QMenu::item { border: 1px solid red; }

Task-number: 252610
Reveiwed-by: jbache