Set the vertical advance of glyph metrics to zero.
The OpenVG paint engine uses this value as the escapement of the glyph
and setting this to the vertical advance results in the text being laid
out in a declining staircase fashion. All other font engines seem to
set the vertical advance to zero so do the same here as well.