Fixed crash in QWizard running on old Windows systems.
When compiling the QWizard on a Windows system where
QT_NO_STYLE_WINDOWSVISTA is not set and running the app
on an old system that has no Vista support, the app would
crash in the paint event of the Vista Back button. In this
scenario, the Vista Back button is not needed (i.e. the
regular Back button next to the Next button is used), so
the fix is simply to avoid instanciating it altogether.

Reviewed-by: janarve
Task-number: 252662