QCss: font-family handle fallback font specs
if one specify more than one parameter in  font-family,
e.g., font-family: Verdana, Arial
Qt should fallback on the second font if the first cannot be found.

QFont::setFamily handle the case when the family name contains a comas,
so we do not need to handle that specially in the css parser code.

Task-number: 252311
Reviewed-by: Thomas Zander