QTemporaryFile: really (re)open file if it has been really closed...
In some circumstances, the file descriptor in QTemporaryFile is actually
closed and setOpenMode alone won't give us reOpen semantics.

Added function to QTemporaryFileEngine that checks if we have open file
handles. On open, if we currently hold no handles, re-open the file.

Trying to open a new file while we hold open handles would lead to
leaks, so added an assert there, to be on the safe side.

Reviewed-by: Thiago