QTemporaryFileEngine now tracks if a fileName has been generated
With recent changes to QTemporaryFile, allowing the file to be closed,
the engine has to keep track of whether a fileName has already been
generated, so we don't generate new files after the first one.

If the file is closed but we already have a name for it, then just
forward the call to the base file engine.

Reviewed-by: Thiago